Fix -Wextra warnings
Compiling with gcc 14 and `-Wextra` flags showed a few instances of invalid
sign compare (comparing int to unsigned long int for ex). So update some
variables to match the types.

To fix sign compare warnings, the return type of `utf8_validate()` changed from
an `int` to `size_t`, and the error result is indicated by `0` and not `-1`, as
before. Previously it returned -1, or 1,2,3,4 values only. Since 0 wasn't used
and it's not a valid return, use it as the new error result.

+++ c_src/jiffy.h
@@ -74,7 +74,7 @@ int int_from_hex(const unsigned char* p);
 int int_to_hex(int val, unsigned char* p);
 int utf8_len(int c);
 int utf8_esc_len(int c);
-int utf8_validate(unsigned char* data, size_t size);
+size_t utf8_validate(unsigned char* data, size_t size);
 int utf8_to_unicode(unsigned char* buf, size_t size);
 int unicode_to_utf8(int c, unsigned char* buf);
 int unicode_from_pair(int hi, int lo);
